+++ CHANGELOG.md
@@ -11,10 +11,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Search function: Display shops matching search terms.
 - Language selector now shows flag color for current language and displays text in uppercase.
+- Translations for player and admin shop indexes
 
 ### Changed
 
 - Already stocked amounts of player shops are subtracted from demands.
+- Separated players and admin shops 
 
 ## [0.6.5] - 2023-12-04